(#316) преобразование типа и yaml config
(OnionGrief/Automatic-Chipollino-Update#1)

+++ libs/InputGenerator/src/TasksGenerator.cpp
@@ -15,7 +15,7 @@ TasksGenerator::TasksGenerator() {
 
 void TasksGenerator::change_seed() {
 	seed_it++;
-	srand((size_t)time(nullptr) + seed_it + rand());
+	srand(static_cast<unsigned int>((size_t)time(nullptr) + seed_it + rand()));
 }
 
 string TasksGenerator::generate_task(int op_num, int max_num_of_func_in_seq_,
